# build/libs/exploded/renren-admin.war/WEB-INF/classes/public/swagger/index.yaml
swagger: '2.0'
info:
  description: 是一套轻量级权限管理系统，极低门槛，拿来即用。设计之初，就非常注重安全性，为企业系统保驾护航，让一切都变得如此简单。
  version: 3.0.0
  title: renren-security
basePath: /renren-admin
schemes:
- http

#定义接口数据
paths:
...
  /sys/role/list:
    get:
      tags:
      - 角色管理
      summary: 角色列表
      produces:
      - application/json
      parameters:
      - name: page
        description: 页码
        in: query
        type: integer
        required: true
      - name: limit
        description: 每页条数
        in: query
        type: integer
        required: true
      - name: sidx
        description: 排序字段
        in: query
        type: string
      - name: order
        description: 排序方式，如：asc、desc
        in: query
        type: string
      - name: roleName
        description: 角色名
        in: query
        type: string
      responses:
        '200':
          description: 返回管理员列表
          schema:
            $ref: '#/definitions/SysRoleEntity'
...
  /sys/role/info/{roleId}:
    get:
      tags:
      - 角色管理
      summary: 获取角色信息
      produces:
      - application/json
      parameters:
      - name: roleId
        description: 角色ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回角色信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /sys/role/save:
    post:
      tags:
      - 角色管理
      summary: 添加角色
      produces:
      - application/json
      parameters:
      - name: body
        description: 角色对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysRoleEntitySave'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/update:
    post:
      tags:
      - 角色管理
      summary: 修改角色
      produces:
      - application/json
      parameters:
      - name: body
        description: 角色对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysRoleEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/role/delete:
    post:
      tags:
      - 角色管理
      summary: 删除角色
      produces:
      - application/json
      parameters:
      - name: body
        description: 角色ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
...
  /sys/menu/info/{menuId}:
    get:
      tags:
      - 菜单管理
      summary: 获取菜单信息
      produces:
      - application/json
      parameters:
      - name: menuId
        description: 菜单ID
        in: path
        type: integer
        required: true
      responses:
        '200':
          description: 返回菜单信息
          schema:
            $ref: '#/definitions/SysRoleEntity'
  /sys/menu/save:
    post:
      tags:
      - 菜单管理
      summary: 添加菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysMenuEntitySave'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/update:
    post:
      tags:
      - 菜单管理
      summary: 修改菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单对象
        in: body
        type: string
        schema:
          $ref: '#/definitions/SysMenuEntityEdit'
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'
  /sys/menu/delete:
    post:
      tags:
      - 菜单管理
      summary: 删除菜单
      produces:
      - application/json
      parameters:
      - name: body
        description: 菜单ID列表
        in: body
        type: array
        items:
          type: integer
          format: int64
          default: 0
        required: true
      responses:
        '200':
          schema:
            $ref: '#/definitions/R'

#  /sys/log/list:
#    get:
#      tags:
#        - 系统日志
#      summary: 日志列表
#      produces:
#        - application/json
#      parameters:
#        - name: page
#          description: 页码
#          in: query
#          type: integer
#          required: true
#        - name: limit
#          description: 每页条数
#          in: query
#          type: integer
#          required: true
#        - name: sidx
#          description: 排序字段
#          in: query
#          type: string
#        - name: order
#          description: 排序方式，如：asc、desc
#          in: query
#          type: string
#        - name: key
#          description: 用户名或用户操作
#          in: query
#          type: string
#      responses:
#        '200':
#          description: 返回日志列表
#          schema:
#            $ref: '#/definitions/SysLogEntity'
#
#  /sys/config/list:
#    get:
#      tags:
#        - 参数管理
#      summary: 参数列表
#      produces:
#        - application/json
#      parameters:
#        - name: page
#          description: 页码
#          in: query
#          type: integer
#          required: true
#        - name: limit
#          description: 每页条数
#          in: query
#          type: integer
#          required: true
#        - name: sidx
#          description: 排序字段
#          in: query
#          type: string
#        - name: order
#          description: 排序方式，如：asc、desc
#          in: query
#          type: string
#        - name: key
#          description: 参数名
#          in: query
#          type: string
#      responses:
#        '200':
#          description: 返回参数列表
#          schema:
#            $ref: '#/definitions/SysConfigEntity'
#  /sys/config/info/{id}:
#    get:
#      tags:
#        - 参数管理
#      summary: 获取参数信息
#      produces:
#        - application/json
#      parameters:
#        - name: id
#          description: 参数ID
#          in: path
#          type: integer
#          required: true
#      responses:
#        '200':
#          description: 返回参数信息
#          schema:
#            $ref: '#/definitions/SysConfigEntity'
#  /sys/config/save:
#    post:
#      tags:
#        - 参数管理
#      summary: 添加参数
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 参数对象
#          in: body
#          type: string
#          schema:
#            $ref: '#/definitions/SysConfigEntity'
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/config/update:
#    post:
#      tags:
#        - 参数管理
#      summary: 修改参数
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 参数对象
#          in: body
#          type: string
#          schema:
#            $ref: '#/definitions/SysConfigEntity'
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/config/delete:
#    post:
#      tags:
#        - 参数管理
#      summary: 删除参数
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 参数ID列表
#          in: body
#          type: array
#          items:
#            type: integer
#            format: int64
#            default: 0
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#
#  /sys/oss/list:
#    get:
#      tags:
#        - 文件服务
#      summary: 文件列表
#      produces:
#        - application/json
#      parameters:
#        - name: page
#          description: 页码
#          in: query
#          type: integer
#          required: true
#        - name: limit
#          description: 每页条数
#          in: query
#          type: integer
#          required: true
#        - name: sidx
#          description: 排序字段
#          in: query
#          type: string
#        - name: order
#          description: 排序方式，如：asc、desc
#          in: query
#          type: string
#      responses:
#        '200':
#          description: 返回文件列表
#          schema:
#            $ref: '#/definitions/SysOssEntity'
#  /sys/oss/config:
#    get:
#      tags:
#        - 文件服务
#      summary: 云存储配置信息
#      produces:
#        - application/json
#      responses:
#        '200':
#          description: 返回云存储配置信息
#          schema:
#            $ref: '#/definitions/SysCloudStorageEntity'
#  /sys/oss/saveConfig:
#    post:
#      tags:
#        - 文件服务
#      summary: 保存云存储配置信息
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 参数对象
#          in: body
#          type: string
#          schema:
#            $ref: '#/definitions/SysCloudStorageEntity'
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/oss/upload:
#    post:
#      tags:
#        - 文件服务
#      summary: 上传文件
#      consumes:
#        - multipart/form-data
#      produces:
#        - application/json
#      parameters:
#        - name: file
#          description: 文件
#          in: formData
#          type: file
#          required: true
#      responses:
#        '200':
#          description: 返回文件列表
#          schema:
#            type: object
#            properties:
#              code:
#                description: 状态码 0：成功  非0：失败
#                type: integer
#                format: int32
#              url:
#                description: 文件URL地址
#                type: string
#              msg:
#                description: 失败原因
#                type: string
#  /sys/oss/delete:
#    post:
#      tags:
#        - 文件服务
#      summary: 删除文件
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 文件ID列表
#          in: body
#          type: array
#          items:
#            type: integer
#            format: int64
#            default: 0
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#
#  /sys/schedule/list:
#    get:
#      tags:
#        - 定时任务
#      summary: 定时任务列表
#      produces:
#        - application/json
#      parameters:
#        - name: page
#          description: 页码
#          in: query
#          type: integer
#          required: true
#        - name: limit
#          description: 每页条数
#          in: query
#          type: integer
#          required: true
#        - name: sidx
#          description: 排序字段
#          in: query
#          type: string
#        - name: order
#          description: 排序方式，如：asc、desc
#          in: query
#          type: string
#        - name: beanName
#          description: spring bean名称
#          in: query
#          type: string
#      responses:
#        '200':
#          description: 返回定时任务列表
#          schema:
#            $ref: '#/definitions/ScheduleJobEntity'
#  /sys/schedule/info/{id}:
#    get:
#      tags:
#        - 定时任务
#      summary: 获取定时任务信息
#      produces:
#        - application/json
#      parameters:
#        - name: jobId
#          description: 定时任务ID
#          in: path
#          type: integer
#          required: true
#      responses:
#        '200':
#          description: 返回定时任务信息
#          schema:
#            $ref: '#/definitions/ScheduleJobEntity'
#  /sys/schedule/save:
#    post:
#      tags:
#        - 定时任务
#      summary: 添加定时任务
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 定时任务对象
#          in: body
#          type: string
#          schema:
#            $ref: '#/definitions/ScheduleJobEntity'
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/schedule/update:
#    post:
#      tags:
#        - 定时任务
#      summary: 修改定时任务
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 定时任务对象
#          in: body
#          type: string
#          schema:
#            $ref: '#/definitions/ScheduleJobEntity'
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/schedule/delete:
#    post:
#      tags:
#        - 定时任务
#      summary: 删除定时任务
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 定时任务ID列表
#          in: body
#          type: array
#          items:
#            type: integer
#            format: int64
#            default: 0
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/schedule/run:
#    post:
#      tags:
#        - 定时任务
#      summary: 立即执行任务
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 定时任务ID列表
#          in: body
#          type: array
#          items:
#            type: integer
#            format: int64
#            default: 0
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/schedule/pause:
#    post:
#      tags:
#        - 定时任务
#      summary: 暂停定时任务
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 定时任务ID列表
#          in: body
#          type: array
#          items:
#            type: integer
#            format: int64
#            default: 0
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'
#  /sys/schedule/resume:
#    post:
#      tags:
#        - 定时任务
#      summary: 恢复定时任务
#      produces:
#        - application/json
#      parameters:
#        - name: body
#          description: 定时任务ID列表
#          in: body
#          type: array
#          items:
#            type: integer
#            format: int64
#            default: 0
#          required: true
#      responses:
#        '200':
#          schema:
#            $ref: '#/definitions/R'

#定义数据模型
definitions:
  R:
    type: object
    properties:
      code:
        description: 状态码 0：成功  非0：失败
        type: integer
        format: int32
      msg:
        description: 失败原因
        type: string
...
  SysRoleEntity:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      #      deptId:
      #        description: 部门ID
      #        type: integer
      #        format: int64
      #      deptName:
      #        description: 部门名称
      #        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
      createUserId:
        description: 创建者ID
        type: integer
        format: int64
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysRoleEntitySave:
    type: object
    properties:
      roleName:
        description: 角色名称
        type: string
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
  SysRoleEntityEdit:
    type: object
    properties:
      roleId:
        description: 角色ID
        type: integer
        format: int64
      roleName:
        description: 角色名称
        type: string
      #      deptId:
      #        description: 部门ID
      #        type: integer
      #        format: int64
      remark:
        description: 备注
        type: string
      menuIdList:
        description: 菜单ID列表
        type: array
        items:
          type: integer
          format: int64
...
  SysMenuEntitySave:
    type: object
    properties:
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
  SysMenuEntityEdit:
    type: object
    properties:
      menuId:
        description: 菜单ID
        type: integer
        format: int64
      name:
        description: 菜单名称
        type: string
      parentId:
        description: 父菜单ID，一级菜单为0
        type: integer
        format: int64
      url:
        description: 菜单URL
        type: string
      perms:
        description: 授权标识
        type: string
      type:
        description: 类型     0：目录   1：菜单   2：按钮
        type: integer
        format: int32
      icon:
        description: 菜单图标
        type: string
      orderNum:
        description: 排序
        type: integer
        format: int32
  SysLogEntity:
    type: object
    properties:
      id:
        description: 日志ID
        type: integer
        format: int64
      username:
        description: 用户名
        type: string
      operation:
        description: 用户操作
        type: string
      method:
        description: 请求方法
        type: string
      params:
        description: 请求参数
        type: string
      time:
        description: 执行时长(毫秒)
        type: integer
        format: int64
      ip:
        description: IP地址
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysConfigEntity:
    type: object
    properties:
      id:
        description: 参数ID
        type: integer
        format: int64
      key:
        description: 参数名
        type: string
      value:
        description: 参数值
        type: string
      remark:
        description: 备注
        type: string
  SysOssEntity:
    type: object
    properties:
      id:
        description: ID
        type: integer
        format: int64
      url:
        description: URL地址
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  SysCloudStorageEntity:
    type: object
    properties:
      type:
        description: 类型 1：七牛  2：阿里云  3：腾讯云
        type: integer
        format: int32
      qiniuDomain:
        description: 七牛绑定的域名
        type: string
      qiniuPrefix:
        description: 七牛路径前缀
        type: string
      qiniuAccessKey:
        description: 七牛ACCESS_KEY
        type: string
      qiniuSecretKey:
        description: 七牛SECRET_KEY
        type: string
      qiniuBucketName:
        description: 七牛存储空间名
        type: string
      aliyunDomain:
        description: 阿里云绑定的域名
        type: string
      aliyunPrefix:
        description: 阿里云路径前缀
        type: string
      aliyunEndPoint:
        description: 阿里云EndPoint
        type: string
      aliyunAccessKeyId:
        description: 阿里云AccessKeyId
        type: string
      aliyunAccessKeySecret:
        description: 阿里云AccessKeySecret
        type: string
      aliyunBucketName:
        description: 阿里云BucketName
        type: string
      qcloudDomain:
        description: 腾讯云绑定的域名
        type: string
      qcloudPrefix:
        description: 腾讯云路径前缀
        type: string
      qcloudAppId:
        description: 腾讯云AppId
        type: string
      qcloudSecretId:
        description: 腾讯云SecretId
        type: string
      qcloudSecretKey:
        description: 腾讯云SecretKey
        type: string
      qcloudBucketName:
        description: 腾讯云BucketName
        type: string
      qcloudRegion:
        description: 腾讯云COS所属地区
        type: string
  ScheduleJobEntity:
    type: object
    properties:
      jobId:
        description: 任务ID
        type: integer
        format: int64
      beanName:
        description: spring bean名称
        type: string
      methodName:
        description: 方法名
        type: string
      params:
        description: 参数
        type: string
      cronExpression:
        description: cron表达式
        type: string
      status:
        description: 任务状态  0：正常  1：暂停
        type: integer
        format: int32
      remark:
        description: 备注
        type: string
      createTime:
        description: 创建时间
        type: string
        format: date-time
  #登录请求对象封装
...